Move VisGraph to components
