+++ resources/index.html
@@ -1,9 +1,22 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
-  <title>React graph vis</title>
+    <meta charset="utf-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1, shrink-to-fit=no">
+    <meta name="theme-color" content="#000000">
+    <title>Audio Reactive LED Strip</title>
   </head>
   <body>
+    <noscript>
+      You need to enable JavaScript to run this app.
+    </noscript>
     <div id="root"></div>
+    <!--
+      This HTML file is a template.
+      If you open it directly in the browser, you will see an empty page.
+
+      You can add webfonts, meta tags, or analytics to this file.
+      The build step will place the bundled scripts into the <body> tag.
+    -->
   <script type="text/javascript" src="/bundle.js"></script></body>
 </html>
