+++ server/package.json
@@ -33,6 +33,8 @@
     "webpack-dev-server": "^3.1.10"
   },
   "dependencies": {
+    "@material-ui/core": "^3.1.0",
+    "@material-ui/icons": "^3.0.1",
     "@types/file-saver": "^2.0.0",
     "file-saver": "^1.3.8",
     "npm": "^6.6.0",
