+++ server/src/App.css
@@ -1,0 +1,7 @@
+
+
+  div#content {
+    display: flex;
+    flex-direction: column;
+    flex: 1;
+  }
