--- server/src/_index.html
@@ -1,152 +1,0 @@
-<!DOCTYPE html>
-<html>
-
-<head>
-  <title>Audio Reactive LED Strip Configuration</title>
-  <style>
-    #network {
-      position: absolute;
-      top: 100px; /* Header Height */
-      bottom: 0px; /* Footer Height */
-      left: 0px;
-      right: 0px;
-      border: 0.5px solid lightgray;
-      background-color:#333333;
-    }
-    #node-popUp {
-      display:none;
-      position:absolute;
-      top:350px;
-      left:170px;
-      z-index:299;
-      width:525px;
-      overflow:auto;
-      background-color: #333333;
-      border-style:solid;
-      border-width:3px;
-      border-color: darkgrey;
-      padding:10px;
-      text-align: center;
-      color: white;
-    }
-    #edge-popUp {
-      display:none;
-      position:absolute;
-      top:350px;
-      left:170px;
-      z-index:299;
-      width:525px;
-      overflow:auto;
-      background-color: #333333;
-      border-style:solid;
-      border-width:3px;
-      border-color: darkgrey;
-      padding:10px;
-      text-align: center;
-      color: white;
-    }
-    #text-area {
-      max-width: 50% !important;
-      min-width: 30% !important;
-    }
-    #node-args {
-      height:50px
-    }
-    div.vis-configuration-wrapper {
-      width: 500px !important;
-    }
-    div.vis-configuration.vis-config-button {
-      display:none !important;
-    }
-    div.vis-configuration.vis-config-option-container {
-      display:none !important;
-    }
-    div.vis-configuration.vis-config-item.vis-config-s2{
-    left:10px;
-    background-color: #333333 !important;
-    padding-left:5px;
-    border-radius:3px;
-    }
-    div.vis-configuration.vis-config-item.vis-config-s3{
-        left:20px;
-        background-color: #333333 !important;
-        padding-left:5px;
-        border-radius:3px;
-    }
-    div.vis-configuration.vis-config-item.vis-config-s4{
-        left:30px;
-        background-color: #333333 !important;
-        padding-left:5px;
-        border-radius:3px;
-    }
-.alert {
-    display:none;
-    padding: 20px;
-    background-color: #f44336;
-    color: white;
-    position:absolute;
-      left: 0px;
-      right: 0px;
-      top: 0px;
-}
-
-.closebtn {
-    margin-left: 15px;
-    color: white;
-    font-weight: bold;
-    float: right;
-    font-size: 22px;
-    line-height: 20px;
-    cursor: pointer;
-    transition: 0.3s;
-}
-
-.closebtn:hover {
-    color: black;
-}
-  </style>
-
-</head>
-<script src="bundle.js" type="text/javascript"></script>
-
-<body>
-  <h1>FilterGraph:</h1>
-  <input type="button" value="save" id="config-saveButton" />
-  load: <input type="file" id="file-input" />
-  <div id="network"></div>
-  <div class="alert" id="alert">
-    <span class="closebtn" onclick="this.parentElement.style.display='none';">&times;</span> 
-    <div id="alert-info"><strong>Danger!</strong> Indicates a dangerous or potentially negative action.</div>
-  </div>
-  <div id="node-popUp">
-    <h2 id="node-operation">node</h2> <br>
-    <div id="node-effectTable">
-      <div class="vis-configuration vis-config-header">effect</div>
-      <div class="vis-configuration vis-config-item vis-config-s2"><select class="form-control" id='node-effectDropdown' name='node-effectDropdown'></select></div>
-    </div>
-    <div id="node-configuration"></div>
-    <table style="margin:auto;">
-      <tr>
-        <td><input type="button" value="save" id="node-saveButton" /></td>
-        <td><input type="button" value="cancel" id="node-cancelButton" /></td>
-      </tr>
-    </table>
-  </div>
-
-
-  <div id="edge-popUp">
-    <h2 id="edge-operation">edge</h2> <br>
-    <div class="vis-configuration vis-config-header">from channel</div>
-    <div class="vis-configuration vis-config-item vis-config-s2"><select class="form-control" id='edge-fromChannelDropdown' name='edge-fromChannelDropdown'></select></div>
-    <div class="vis-configuration vis-config-header">to channel</div>
-    <div class="vis-configuration vis-config-item vis-config-s2"><select class="form-control" id='edge-toChannelDropdown' name='edge-toChannelDropdown'></select></div>
-    <table style="margin:auto;">
-      <tr>
-        <td><input type="button" value="save" id="edge-saveButton" /></td>
-        <td><input type="button" value="cancel" id="edge-cancelButton" /></td>
-      </tr>
-    </table>
-  </div>
-
-</body>
-</html>
