--- server/src/index.css
@@ -1,106 +1,0 @@
-.vis-network {
-  border: 0.5px solid lightgray;
-  background-color: #333333;
-}
-
-#node-popUp {
-  display: none;
-  position: absolute;
-  top: 350px;
-  left: 170px;
-  z-index: 299;
-  width: 525px;
-  overflow: auto;
-  background-color: #333333;
-  border-style: solid;
-  border-width: 3px;
-  border-color: darkgrey;
-  padding: 10px;
-  text-align: center;
-  color: white;
-}
-
-#edge-popUp {
-  display: none;
-  position: absolute;
-  top: 350px;
-  left: 170px;
-  z-index: 299;
-  width: 525px;
-  overflow: auto;
-  background-color: #333333;
-  border-style: solid;
-  border-width: 3px;
-  border-color: darkgrey;
-  padding: 10px;
-  text-align: center;
-  color: white;
-}
-
-#text-area {
-  max-width: 50% !important;
-  min-width: 30% !important;
-}
-
-#node-args {
-  height: 50px
-}
-
-div.vis-configuration-wrapper {
-  width: 500px !important;
-}
-
-div.vis-configuration.vis-config-button {
-  display: none !important;
-}
-
-div.vis-configuration.vis-config-option-container {
-  display: none !important;
-}
-
-div.vis-configuration.vis-config-item.vis-config-s2 {
-  left: 10px;
-  background-color: #333333 !important;
-  padding-left: 5px;
-  border-radius: 3px;
-}
-
-div.vis-configuration.vis-config-item.vis-config-s3 {
-  left: 20px;
-  background-color: #333333 !important;
-  padding-left: 5px;
-  border-radius: 3px;
-}
-
-div.vis-configuration.vis-config-item.vis-config-s4 {
-  left: 30px;
-  background-color: #333333 !important;
-  padding-left: 5px;
-  border-radius: 3px;
-}
-
-.alert {
-  display: none;
-  padding: 20px;
-  background-color: #f44336;
-  color: white;
-  position: absolute;
-  left: 0px;
-  right: 0px;
-  top: 0px;
-}
-
-.closebtn {
-  margin-left: 15px;
-  color: white;
-  font-weight: bold;
-  float: right;
-  font-size: 22px;
-  line-height: 20px;
-  cursor: pointer;
-  transition: 0.3s;
-}
-
-.closebtn:hover {
-  color: black;
-}
